separate progress-ring specific exports and use stub modules
