+++ packages/web-components/fast-foundation/src/progress-ring/index.ts
@@ -1,1 +1,3 @@
+export * from "./progress-ring.js";
+export * from "./progress-ring.options.js";
 export * from "./progress-ring.template.js";
