+++ packages/web-components/fast-foundation/src/progress-ring/progress-ring.options.ts
@@ -1,0 +1,10 @@
+import type { SyntheticViewTemplate } from "@microsoft/fast-element";
+import type { FoundationElementDefinition } from "../index.js";
+
+/**
+ * ProgressRing configuration options
+ * @public
+ */
+export type ProgressRingOptions = FoundationElementDefinition & {
+    indeterminateIndicator?: string | SyntheticViewTemplate;
+};
