+++ packages/web-components/fast-foundation/src/progress-ring/progress-ring.ts
@@ -1,0 +1,1 @@
+export { BaseProgress as ProgressRing } from "../progress/base-progress.js";
