+++ packages/web-components/fast-foundation/src/progress/base-progress.ts
@@ -1,30 +1,6 @@
-import {
-    attr,
-    nullableNumberConverter,
-    observable,
-    SyntheticViewTemplate,
-} from "@microsoft/fast-element";
-import {
-    FoundationElement,
-    FoundationElementDefinition,
-} from "../foundation-element/foundation-element.js";
+import { attr, nullableNumberConverter, observable } from "@microsoft/fast-element";
+import { FoundationElement } from "../foundation-element/foundation-element.js";
 
-/**
- * Progress configuration options
- * @public
- */
-export type ProgressOptions = FoundationElementDefinition & {
-    indeterminateIndicator1?: string | SyntheticViewTemplate;
-    indeterminateIndicator2?: string | SyntheticViewTemplate;
-};
-
-/**
- * ProgressRing configuration options
- * @public
- */
-export type ProgressRingOptions = FoundationElementDefinition & {
-    indeterminateIndicator?: string | SyntheticViewTemplate;
-};
 /**
  * An Progress HTML Element.
  * Implements the {@link https://www.w3.org/TR/wai-aria-1.1/#progressbar | ARIA progressbar }.
