+++ packages/web-components/fast-foundation/src/progress/index.ts
@@ -1,2 +1,3 @@
 export * from "./base-progress.js";
+export * from "./progress.options.js";
 export * from "./progress.template.js";
