+++ packages/web-components/fast-foundation/src/progress/progress.options.ts
@@ -1,0 +1,11 @@
+import type { SyntheticViewTemplate } from "@microsoft/fast-element";
+import type { FoundationElementDefinition } from "../index.js";
+
+/**
+ * Progress configuration options
+ * @public
+ */
+export type ProgressOptions = FoundationElementDefinition & {
+    indeterminateIndicator1?: string | SyntheticViewTemplate;
+    indeterminateIndicator2?: string | SyntheticViewTemplate;
+};
