+++ packages/web-components/fast-foundation/src/progress/progress.template.ts
@@ -1,7 +1,8 @@
-import { html, when } from "@microsoft/fast-element";
 import type { ViewTemplate } from "@microsoft/fast-element";
+import { html, when } from "@microsoft/fast-element";
 import type { FoundationElementTemplate } from "../foundation-element/foundation-element.js";
-import type { BaseProgress, ProgressOptions } from "./base-progress.js";
+import type { BaseProgress } from "./base-progress.js";
+import type { ProgressOptions } from "./progress.options.js";
 
 /**
  * The template for the {@link @microsoft/fast-foundation#BaseProgress} component.
